updated wait for deployments location
